flask-project/requirements.txt
absl-py==0.12.0
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.7.0
astroid==2.5.6
astunparse==1.6.3
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.7
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorhash==1.0.3
configparser==5.0.2
constantly==15.1.0
cryptography==3.4.6
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dateparser==1.0.0
defusedxml==0.7.1
fastapi==0.63.0
feedparser==6.0.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Compress==1.9.0
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-MonitoringDashboard==3.1.0
flask-restx==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.5.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.2.0
plotly==4.14.3
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.1
protobuf==3.15.7
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==1.4.2
pylint==2.7.2
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.7.2
scipy==1.6.3
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zeep==4.0.0
zope.interface==5.2.0
